Fixed tests to be more deterministic and avoid leaving directories that can't be deleted in Explorer

+++ Tests/DirectoryTests.cs
@@ -20,6 +20,7 @@ namespace Tests
 		private static string rootTestDir;
 		private static string longPathDirectory;
 		private static string longPathFilename;
+		private static string longPathRoot;
 		private const string Filename = "filename.ext";
 
 		[ClassInitialize]
@@ -27,6 +28,7 @@ namespace Tests
 		{
 			rootTestDir = context.TestDir;
 			longPathDirectory = Util.MakeLongPath(rootTestDir);
+			longPathRoot = longPathDirectory.Substring(0, context.TestDir.Length + 1 + longPathDirectory.Substring(rootTestDir.Length + 1).IndexOf('\\'));
 			Directory.CreateDirectory(longPathDirectory);
 			Debug.Assert(Directory.Exists(longPathDirectory));
 			longPathFilename = new StringBuilder(longPathDirectory).Append(@"\").Append(Filename).ToString();
@@ -485,8 +487,8 @@ namespace Tests
 
 			const bool recursive = true;
 			Directory.Delete(tempLongPathFilename1, recursive);
+			Directory.Delete(tempLongPathFilename2, recursive);
 		}
-
 
 		[TestMethod]
 		public void TestGetAccessControl()
@@ -1012,7 +1014,8 @@ namespace Tests
 		{
 			try
 			{
-				File.Delete(longPathFilename);
+				if (File.Exists(longPathFilename))
+					File.Delete(longPathFilename);
 			}
 			catch (Exception e)
 			{
@@ -1021,7 +1024,8 @@ namespace Tests
 			}
 			finally
 			{
-				Directory.Delete(longPathDirectory, true);
+				if (Directory.Exists(longPathRoot))
+					Directory.Delete(longPathRoot, true);
 			}
 		}
 	}
