+++ Tests/FileTests.cs
@@ -55,6 +55,7 @@ namespace Tests
 		{
 			rootTestDir = context.TestDir;
 			longPathDirectory = Util.MakeLongPath(rootTestDir);
+			longPathRoot = longPathDirectory.Substring(0, context.TestDir.Length + 1 + longPathDirectory.Substring(rootTestDir.Length + 1).IndexOf('\\'));
 			Directory.CreateDirectory(longPathDirectory);
 			Debug.Assert(Directory.Exists(longPathDirectory));
 			longPathFilename = new StringBuilder(longPathDirectory).Append(@"\").Append(Filename).ToString();
@@ -1303,7 +1304,8 @@ namespace Tests
 				File.Delete(filename);
 			}
 		}
-			string filename = Util.CreateNewFile(longPathDirectory);
+		
+		private static string longPathRoot;
 
 		/// <remarks>
 		/// TODO: more realistic FileSecurity scenarios
@@ -1332,7 +1334,8 @@ namespace Tests
 		{
 			try
 			{
-				File.Delete(longPathFilename);
+				if (File.Exists(longPathFilename))
+					File.Delete(longPathFilename);
 			}
 			catch (Exception e)
 			{
@@ -1341,7 +1344,8 @@ namespace Tests
 			}
 			finally
 			{
-				Directory.Delete(longPathDirectory, true);
+				if(Directory.Exists(longPathRoot))
+					Directory.Delete(longPathRoot, true);
 			}
 		}
 	}
