+++ Tests/PathTests.cs
@@ -25,6 +25,7 @@ namespace Tests
 		private static string rootTestDir;
 		private static string longPathDirectory;
 		private static string longPathFilename;
+		private static string longPathRoot;
 		private const string Filename = "filename.ext";
 
 		[ClassInitialize]
@@ -32,6 +33,7 @@ namespace Tests
 		{
 			rootTestDir = context.TestDir;
 			longPathDirectory = Util.MakeLongPath(rootTestDir);
+			longPathRoot = longPathDirectory.Substring(0, context.TestDir.Length + 1 + longPathDirectory.Substring(rootTestDir.Length + 1).IndexOf('\\'));
 			Directory.CreateDirectory(longPathDirectory);
 			Debug.Assert(Directory.Exists(longPathDirectory));
 			longPathFilename = new StringBuilder(longPathDirectory).Append(@"\").Append(Filename).ToString();
@@ -409,7 +411,8 @@ namespace Tests
 		{
 			try
 			{
-				File.Delete(longPathFilename);
+				if (File.Exists(longPathFilename))
+					File.Delete(longPathFilename);
 			}
 			catch (Exception e)
 			{
@@ -418,7 +421,8 @@ namespace Tests
 			}
 			finally
 			{
-				Directory.Delete(longPathDirectory, true);
+				if(Directory.Exists(longPathRoot))
+					Directory.Delete(longPathRoot, true);
 			}
 		}
 	}
